lab-01-end-to-end-training/requirements.txt
pandas
matplotlib
seaborn
pillow
openpyxl
numpy
torch
torchvision
